feat: 참가 신청 알림

+++ backend/booking/src/main/java/com/booking/booking/global/dto/request/EnrollNotificationRequest.java
@@ -1,7 +1,7 @@
 package com.booking.booking.global.dto.request;
 
 public record EnrollNotificationRequest(
-        Long memberId,
+        Integer memberId,
         String meetingTitle
 ) {
 }
